requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.7
